Added JSON verification back for the detection settings
